Major Merge to upgrade architect

+++ src/main/java/com/techcrack/bookwise/dtos/borrowbook/request/ReturnBookRequest.java
@@ -2,7 +2,7 @@ package com.techcrack.bookwise.dtos.borrowbook.request;
 
 import com.techcrack.bookwise.dtos.borrowbook.layer.ReturnBookContext;
 
-public record ReturnBookRequest(long borrowBookId, long amount) {
+public record ReturnBookRequest(long borrowBookId, double amount) {
     public ReturnBookContext buildContext(long userId) {
         return new ReturnBookContext(borrowBookId, amount);
     }
